% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform angle units.R
\name{angleTransform}
\alias{angleTransform}
\title{Transform angle units}
\usage{
angleTransform(x, type = "d")
}
\arguments{
\item{x}{numeric. The angle value}

\item{type}{character. The unit of the angle, can be "d", "deg","degree" for degrees or "r","rad","radian" for radians. default: "d"}
}
\value{
numeric. The angle converted to the corresponding unit
}
\description{
Convert angle units from degrees to radians or viceversa
}
\details{
Transform angle units

This function allows to convert angle units from degrees to radians or viceversa, by default it assumes that the input is in degrees
}
\note{
be aware of the type of unit used
}
\examples{
angleTransform(180)  # 3.141593
angleTransform(3.141593, "r") #180
angleTransform(270, "d") #4.712388
}
\references{
"Trigonometry" by Jane Smith
}
