% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Atmospheric pressure.R
\name{atmPressure}
\alias{atmPressure}
\title{Calculates atmospheric pressure in kilopascals (kPa) based on altitude and selected method}
\usage{
atmPressure(alt, method = "isa")
}
\arguments{
\item{alt}{numeric, the altitude in meters above sea level.}

\item{method}{character, the method to be used for calculating atmospheric pressure. Default is "isa" and allows for selection between "isa" and "equation".}
}
\value{
The function returns a number in kilopascals (kPa) that corresponds to the atmospheric pressure based on the input altitude and selected method.
}
\description{
The "atmPressure()" function allows for the calculation of atmospheric pressure in kilopascals (kPa) based on the input altitude and selected method. Two methods can be selected: the International Standard Atmosphere (ISA) model and a simplified equation based on the ideal gas law at a standard temperature of 20°C.
}
\note{
It is important to note that the simplified equation based on the ideal gas law is an approximation and does not take into account factors such as humidity and air pollution, so its accuracy may vary in real-world conditions.
}
\examples{
atmPressure(500)
# Result: 95.46078
atmPressure(500, "equation")
# Result: 95.52765
}
\references{
\itemize{
\item International Civil Aviation Organization. (2019). "International Standard Atmosphere (ISA)". Retrieved from https://www.icao.int/publications/DOC7488/Pages/Search.aspx
\item Halliday, D., Resnick, R., & Walker, J. (2018). Fundamentals of Physics. John Wiley & Sons.
\item Bailyn, M. (2004). Introduction to Thermodynamics and Kinetic Theory of Gases. Cambridge University Press.
}
}
