% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Real vapor pressure.R
\name{ea}
\alias{ea}
\title{Actual vapor pressure}
\usage{
ea(relh, tmin, tmax)
}
\arguments{
\item{relh}{relative humidity in \%.}

\item{tmin}{temperature minimum in °C.}

\item{tmax}{temperature maximum in °C.}

\item{tavg}{temperature average in °C.}
}
\value{
numeric value of actual vapor pressure in hPa.
}
\description{
This function calculates the actual vapor pressure (ea) using three different methods.
\describe{
\item{Method 1}{If only relative humidity (relh) and temperature average (tavg) are provided, ea is calculated as relh * es(tavg) / 100, where es(tavg) is the saturation vapor pressure calculated using the es() function.}
\item{Method 2}{If only temperature minimum (tmin) is provided, the saturation vapor pressure (es) at tmin is returned.}
\item{Method 3}{If relative humidity (relh), temperature minimum (tmin) and temperature maximum (tmax) are provided, ea is calculated as relh * mean(es(tmin), es(tmax)) / 100}
}
}
\note{
A temperature minimum (tmin) approximation to dew point temperature is used when relative humidity is not provided.
}
\examples{
ea(relh = 60, tavg = 25)
ea(tmin = 5)
ea(relh = 60, tmin = 5, tmax = 10)

}
\references{
\itemize{
\item Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). Crop evapotranspiration-guidelines for computing crop water requirements-FAO irrigation and drainage paper 56. Rome, Italy: Food and Agriculture Organization of the United Nations (FAO).
\item Richardson, L. (1922). The Approximation of the Dew-Point to the Minimum Temperature. Monthly Weather Review, 50(8), 637-638.
\item FAO 56. Penman-Monteith equation for evapotranspiration. FAO Irrigation and drainage paper 56. (https://www.fao.org/3/X0490E/x0490e00.htm)
\item Monteith, J. L. (1965).Evaporation and environment. Symp. Soc. Exp. Biol., 19, 205-234.
}
}
