% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{gamma}
\alias{gamma}
\title{constante psicrométrica gamma}
\usage{
gamma(x, lambda = 2.45, cp = 1.013 * 0.001, epsilon = 0.622)
}
\arguments{
\item{x}{presion kPa}

\item{lambda, cp, epsilon}{is contants}
}
\value{
a numeric value
}
\description{
Obtiene gamma a partir de la presion atmosférica
}
