% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{get.delta}
\alias{get.delta}
\title{Pendiente de la curva de presion de vapor}
\usage{
get.delta(ta, es = NULL)
}
\arguments{
\item{ta}{temperature ºC}

\item{es}{temperature ºC}
}
\description{
Obtiene delta a partir de la temperatura
}
