% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Real vapor pressure.R
\name{get.ea}
\alias{get.ea}
\title{Real vapor pressure}
\usage{
get.ea(tavg = NULL, tmin = NULL, tmax = NULL, relh = NULL)
}
\arguments{
\item{tavg}{temperature ºC}

\item{relh}{relative humidity \%}
}
\value{
a numeric value kPa
}
\description{
Obtiene la presion de vapor real a partir de la temperatura
}
