% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{get.lambda}
\alias{get.lambda}
\title{calor latente de vaporización}
\usage{
get.lambda(ta)
}
\arguments{
\item{ta}{temperature ºC}
}
\value{
a numeric value kPa
}
\description{
Obtiene lambda a partir de la temperatura
}
