% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Net shortwave radiation.R
\name{get.Rns}
\alias{get.Rns}
\title{Radiación neta de onda corta}
\usage{
get.Rns(srad, albedo = NULL)
}
\arguments{
\item{srad}{radiación solar}

\item{albedo}{opcional}
}
\value{
a numeric value
}
\description{
Obtiene la radiación neta de onda corta
}
