% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Net shortwave radiation.R
\name{Rns}
\alias{Rns}
\title{Net shortwave radiation}
\usage{
Rns(srad, albedo = NULL)
}
\arguments{
\item{srad}{numeric, solar radiation (W/m^2)}

\item{albedo}{numeric, albedo of the surface (dimensionless), default value is NULL}
}
\value{
numeric, net shortwave radiation (W/m^2).
}
\description{
The function calculates the net shortwave radiation given the solar radiation (srad) and albedo. If albedo is not given, it assumes a default value of 0.77
}
\note{
The value of albedo should be between 0 and 1
}
\examples{
Rns(srad = 32)
Rns(srad = 32, albedo = 0.25)
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). Crop evapotranspiration-guidelines for computing crop water requirements-FAO irrigation and drainage paper 56. Rome, Italy: Food and Agriculture Organization of the United Nations (FAO).
}
