% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clear sky solar irradiation.R
\name{Rso}
\alias{Rso}
\title{Clear sky solar irradiation (Rso)}
\usage{
Rso(Ra, alt)
}
\arguments{
\item{Ra}{Extraterrestrial solar radiation \link{MJ m^-2 day^-1}}

\item{alt}{Altitude \link{m}}
}
\value{
Rso Clear sky solar irradiation \link{MJ m^-2 day^-1}
}
\description{
The function calculates the clear sky solar irradiation (Rso) from extraterrestrial solar radiation (Ra) and the altitude (alt) as described in the FAO 56 Penman-Monteith equation for evapotranspiration.
}
\note{
The formula used in this function is Rso=Ra*(0.75+alt*2e-5)
}
\examples{
Rso(Ra = 13.5, alt = 1000)
# Output: 9.9625
}
\references{
FAO. (1998). Irrigation and Drainage Paper 56: Crop evapotranspiration - Guidelines for computing crop water requirements - FAO Irrigation and Drainage Paper 56. FAO.
}
