% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solar Declination.R
\name{solarDeclination}
\alias{solarDeclination}
\title{Solar Declination}
\usage{
solarDeclination(i)
}
\arguments{
\item{i}{numeric. The day of the year, between 1 and 365}
}
\value{
numeric value, the solar declination angle in radians
}
\description{
This function calculates the solar declination angle based on the day of the year
}
\details{
Solar Declination
}
\note{
the input day of year should be an integer between 1 and 365
}
\examples{
solarDeclination(90)  # 0.06315547
solarDeclination(270) # -0.04569977
}
\references{
"Astronomy Algorithms" by Jean Meeus
}
