% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extraterrestrial solar radiation.R
\name{SolarIrradiance}
\alias{SolarIrradiance}
\title{Extraterrestrial solar radiation or radiation TOA}
\usage{
SolarIrradiance(latitude, months = NULL, dates = NULL, doy = NULL)
}
\arguments{
\item{latitude}{geographic coordinate specifying the north-south position
of a point on the Earth's surface in sexagecimal degrees}
}
\value{
a numeric value in kJ/m2
}
\description{
This function
}
