% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extraterrestrial solar radiation.R
\name{SolarIrradiance}
\alias{SolarIrradiance}
\title{Extraterrestrial Solar Radiation or Radiation TOA}
\usage{
SolarIrradiance(latitude, months = NULL, dates = NULL, doy = NULL)
}
\arguments{
\item{latitude}{Numeric value specifying the geographic coordinate of the north-south position
of a point on the Earth's surface in sexagesimal degrees.}

\item{months}{Optional, a numeric vector representing the month(s) for which to calculate solar irradiance.}

\item{dates}{Optional, a vector of dates (Date objects) for which to calculate solar irradiance.}

\item{doy}{Optional, a numeric vector representing the day(s) of the year for which to calculate solar irradiance.}
}
\value{
A numeric value representing extraterrestrial solar radiation in kJ/m^2.
}
\description{
This function calculates the extraterrestrial solar radiation or radiation
at the top of the atmosphere (TOA) given the latitude and either the month, date, or day of the year.
}
\note{
This function makes some simplifications in calculating extraterrestrial solar radiation and may not be applicable in all cases. The input latitude is assumed to be in decimal degrees, and the output radiation is in kJ/m^2. For further analysis of solar radiation, consider using other functions in the spAgroClimate package. Please note that this function is valid for latitude values between -66.9 and 66.9 degrees.
}
\examples{
SolarIrradiance(latitude = 45, doy = 172) # June 21st, using day of the year
SolarIrradiance(latitude = 45, months = 6) # June, using month
SolarIrradiance(latitude = 45, dates = as.Date("2015-06-21")) # June 21st, using date
}
\references{
Duffie, J. A., & Beckman, W. A. (2013). Solar engineering of thermal processes. John Wiley & Sons.

Iqbal, M. (2012). An introduction to solar radiation. Elsevier.

Liou, K. N. (2002). An introduction to atmospheric radiation. Academic Press.

Liu, B. Y., & Jordan, R. C. (1960). The interrelationship and characteristic distribution of direct, diffuse and total solar radiation. Solar Energy, 4(3), 1-19.
}
